Run/Debug toolbar buttons (#407).
Fixes: #407.
