+++ resources/META-INF/plugin.xml
@@ -99,11 +99,11 @@
   <!-- See: http://www.jetbrains.org/intellij/sdk/docs/basics/getting_started/plugin_compatibility.html -->
 
   <actions>
-    <action id="Flutter.HotReloadFlutterAppKey" class="io.flutter.actions.HotReloadFlutterAppKeyAction" text="Flutter Hot Reload">
+    <action id="Flutter.HotReloadFlutterAppKey" class="io.flutter.actions.HotReloadFlutterAppRetarget" text="Flutter Hot Reload">
       <keyboard-shortcut keymap="$default" first-keystroke="ctrl alt SEMICOLON"/>
       <keyboard-shortcut keymap="$default" first-keystroke="ctrl BACK_SLASH"/>
     </action>
-    <action id="Flutter.RestartFlutterAppKey" class="io.flutter.actions.RestartFlutterAppKeyAction" text="Flutter Restart Application">
+    <action id="Flutter.RestartFlutterAppKey" class="io.flutter.actions.RestartFlutterAppRetarget" text="Flutter Restart Application">
       <keyboard-shortcut keymap="$default" first-keystroke="shift ctrl alt SEMICOLON"/>
       <keyboard-shortcut keymap="$default" first-keystroke="shift ctrl BACK_SLASH"/>
     </action>
@@ -157,6 +157,20 @@
       <separator/>
       <add-to-group group-id="ProjectViewPopupMenu" relative-to-action="CutCopyPasteGroup" anchor="before"/>
      </group>
+
+    <!-- main toolbar run actions -->
+    <group id="Flutter.MainToolbarActions.Run">
+      <separator/>
+      <action id="Flutter.Toolbar.ReloadAction" class="io.flutter.actions.HotReloadFlutterAppRetarget"
+              description="Reload"
+              icon="FlutterIcons.ReloadBoth"/>
+      <action id="Flutter.Toolbar.RestartAction" class="io.flutter.actions.RestartFlutterAppRetarget"
+              description="Restart"
+              icon="FlutterIcons.Restart"/>
+      <separator/>
+      <add-to-group anchor="last" group-id="RunContextGroup" />
+      <add-to-group anchor="last" group-id="ToolbarRunGroup" />
+    </group>
 
   </actions>
