--- src/io/flutter/actions/RestartFlutterAppKeyAction.java
@@ -1,16 +1,0 @@
-/*
- * Copyright 2016 The Chromium Authors. All rights reserved.
- * Use of this source code is governed by a BSD-style license that can be
- * found in the LICENSE file.
- */
-package io.flutter.actions;
-
-/**
- * A keystroke invoked {@link RestartFlutterApp} action.
- */
-public class RestartFlutterAppKeyAction extends FlutterKeyAction {
-
-  public RestartFlutterAppKeyAction() {
-    super(RestartFlutterApp.ID);
-  }
-}
